+++ imdb/tensorflow/keras/requirements.txt
@@ -1,3 +1,3 @@
-tensorflow==1.10.*
+tensorflow==1.12.*
 requests==2.19.1
 setuptools==39.1.0
